feature:策略概率装配处理

+++ big-market-infrastructure/src/main/java/com/big/market/infrastructure/infrastructure/persistent/po/Award.java
@@ -14,7 +14,7 @@ public class Award {
     /** 自增ID */
     private Long id;
     /** 抽奖奖品ID - 内部流转使用 */
-    private Long awardId;
+    private Integer awardId;
     /** 奖品对接标识 - 每一个都是一个对应的发奖策略 */
     private String awardKey;
     /** 奖品配置信息 */
