+++ big-market-infrastructure/src/main/java/com/big/market/infrastructure/infrastructure/persistent/po/StrategyRule.java
@@ -17,7 +17,7 @@ public class StrategyRule {
     /** 抽奖策略ID */
     private Long strategyId;
     /** 抽奖奖品ID【规则类型为策略，则不需要奖品ID】 */
-    private Long awardId;
+    private Integer awardId;
     /** 抽象规则类型；1-策略规则、2-奖品规则 */
     private Integer ruleType;
     /** 抽奖规则类型【rule_random - 随机值计算、rule_lock - 抽奖几次后解锁、rule_luck_award - 幸运奖(兜底奖品)】 */
